Enforce line length rule on CI
Our style guide insists on lines no longer than 100 characters, but we
actually don't enforce that rule and regressed it quite a bit. We have a
checker for that in tools/lint.zig, but we hardly ever run it.

Fix the issue by moving the check into a proper test, which would be
guaranteed to run by CI.

As we have a lot of violations already, and don't want to fix them all
in one go, this PR implements a ratchet system with a list of known
exceptions.